requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
certifi>=2024.0.0
tenacity>=8.2.0
fastapi>=0.135,<1
uvicorn>=0.42,<1
slowapi>=0.1.9
